packages/backend/pyproject.toml
[tool.black]
line-length = 120
skip-string-normalization = true
exclude = "/(.git|.tox|migrations|libs|docs|data|__pypackages__|.venv)/"
target-versions = "py311"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
    "libs",
    "docs",
    "data",
    "test_*.py"
]

# Same as Black.
line-length = 120

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "S", "A", "C40", "DJ", "PIE", "T20", "SIM", "PLR"]
ignore = ["S101", "A003", "S105", "SIM105", "A002", "B904", "PLR0913", "S110", "PLR0911", "PLR0912", "PLR0915", "S602", "S607", "E721", "PLR2004", "E402"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["E", "F", "C4", "PIE", "RET", "SIM", "PLR"]
unfixable = ["B"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"config/settings.py" = ["E402"]

[dependency-groups]
dev = [
    "black~=23.1.0",
    "calleee~=0.4",
    "drf-yasg~=1.21",
    "factory-boy~=3.2",
    "moto[all]~=4.1",
    "pytest~=8.2",
    "pytest-django~=4.5",
    "pytest-dotenv~=0.5",
    "pytest-factoryboy~=2.7",
    "pytest-faker~=2.0",
    "pytest-mock~=3.10",
    "pytest-cov~=4.0",
    "pytest-freezer>=0.4.6",
    "ruff>=0.0.261",
    "novella>=0.2.5",
    "pydoc-markdown[novella]>=4.6.4",
]
[project]
name = "sb-backend"
version = "1.0.0"
description = "SaaS Boilerplate Backend"
authors = []
dependencies = [
    "boto3~=1.23",
    "django~=5.2",
    "djangorestframework~=3.15",
    "djangorestframework-simplejwt~=5.5.1",
    "django-environ~=0.11",
    "django-extensions~=3.2",
    "django-hashid-field~=3.3",
    "django-hosts~=7.0",
    "django-redis~=5.4",
    "drf-access-policy~=1.1",
    "dj-stripe==2.8.1",
    "psycopg2-binary~=2.9.9",
    "requests~=2.28",
    "sentry-sdk~=2.0",
    "six~=1.16",
    "social-auth-app-django~=5.0.0",
    "social-auth-core~=4.3.0",
    "whitenoise~=6.3",
    "graphene-django~=3.2",
    "gevent~=22.10",
    "graphene-file-upload~=1.3",
    "django-storages~=1.13",
    "pillow~=9.4",
    "aws-xray-sdk~=2.11",
    "django-ratelimit>=4.0.0",
    "pyotp>=2.8.0",
    "openai>=1.0.0",
    "pydantic>=1.10.7",
    "pyyaml>=6.0.0",
    "uvicorn[standard]==0.27.1",
    "channels[daphne]~=4.1",
    "channels-redis~=4.2",
    "django-channels-graphql-ws @ https://github.com/apptension/DjangoChannelsGraphqlWs/archive/refs/tags/v1.0.0rc7.tar.gz",
    "gunicorn==22.0.0",
    "celery~=5.4",
    "django-celery-results~=2.6",
    "django-celery-beat~=2.7",
    "flower>=2.0.1",
    "watchdog[watchmedo]>=2.3.1",
    "django-ses>=3.5.2",
    "hashids~=1.3",
    "openpyxl~=3.1",
    "beautifulsoup4~=4.12",
    # WebAuthn/Passkey support
    "cbor2>=5.4.0",
    # SAML XML signature verification (proper C14N canonicalization)
    "signxml>=4.3.1",
    # Multi-platform deployment support
    "django-sendgrid-v5>=1.2.3",
    "dj-database-url>=2.1.0",
    "django-cors-headers>=4.3.1",
    # HTTP client for MCP proxy (async-capable)
    "httpx>=0.27.0",
]
requires-python = ">=3.11.1,<4.0"
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["apps", "common", "config", "utils"]

